Don't render empty orders
