+++ client/src/components/ShoppingCart.js
@@ -1,7 +1,7 @@
 import Button from "react-bootstrap/Button"
 import Row from "react-bootstrap/Row"
 import Col from "react-bootstrap/Col"
-import { submitOrders } from "../reducers/shoppingCart"
+import { submitOrders } from "../actions/shoppingCart"
 import { useDispatch, useSelector } from "react-redux"
 
 const ShoppingCart = () => {
@@ -31,24 +31,32 @@ const ShoppingCart = () => {
     }
   }
 
+  const orderHasSizes = (order) => {
+    return order.batches.reduce((acc, batch) => acc + batch.order_quantity, 0) > 0
+  }
+
   return (
     <Row className="mt-5 mx-2">
       <Col xs={{ span: 8, offset: 2 }} className="mb-4 text-center">
         <h2 className="mb-4">Ostoskori</h2>
-        {cart.map((order) => {
-          return (
-            <div>
-              Noutotilaus paikassa {order.event.name} <br /> Tuotteet: <br />
-              {order.batches.map((batch) => {
-                return (
-                  <p>
-                    {batch.order_quantity} x {batch.product.name} {batch.unit}{" "}
-                    {batch.product.type}
-                  </p>
-                )
-              })}
-            </div>
-          )
+        {cart.map((order, index) => {
+          if (orderHasSizes(order)) {
+            return (
+              <div key={index}>
+                Noutotilaus paikassa {order.event.name} <br /> Tuotteet: <br />
+                {order.batches.map((batch) => {
+                  if (batch.order_quantity > 0) {
+                    return (
+                      <p>
+                        {batch.order_quantity} x {batch.product.name} {batch.unit}{" "}
+                        {batch.product.type}
+                      </p>
+                    )
+                  } else return null
+                })}
+              </div>
+            )
+          } else return null
         })}
         <Button variant="success" onClick={handleSubmitOrders}>
           Lähetä tilaus
